% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{colmat2lutfile}
\alias{colmat2lutfile}
\title{Write color matrix to LUT file}
\usage{
colmat2lutfile(
  colmat,
  file,
  fill = TRUE,
  length = 256,
  bw = TRUE,
  truncate = FALSE,
  ...
)
}
\arguments{
\item{colmat}{color matrix}

\item{file}{file to write to}

\item{fill}{whether to fill}

\item{length}{desired length (only when \code{fill} is TRUE)}

\item{bw}{add 'black' as first and 'white' as last entry?}

\item{truncate}{truncate if too long}

\item{...}{ignored}
}
\value{
color matrix that was written (possibly filled)
}
\description{
Companion function to \code{\link{writelut}} that sets up a correct
lookup table.
}
\examples{
colmat2lutfile(rainbow_lab(6), tempfile(), fill = FALSE)
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
