% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{color.shades}
\alias{color.shades}
\title{create color shades}
\usage{
color.shades(
  reps,
  col = col.fun(length(reps)),
  scale = ifelse(direction == "bright", 0.7, 0.5),
  gscale = TRUE,
  direction = c("all", "bright", "dark"),
  col.fun = default.hcl,
  ...
)
}
\arguments{
\item{reps}{vector of repetitions}

\item{col}{color vector; must be a valid argument to \code{\link[grDevices]{col2rgb}}
or a \eqn{3\times N}{3xN} RGB color matrix (one column per color).
see description}

\item{scale}{how much of the range to black/white should be covered}

\item{gscale}{global shade steps?}

\item{direction}{which way to build the shades, see Description}

\item{col.fun}{function to calculate default colors}

\item{...}{passed to \code{\link{ColorShadeRamp}}}
}
\value{
Color matrix with on entry per column
}
\description{
For each color in \code{col} create as many shades as indicated by the value
of \code{reps} at the corresponding index by grading towards 'white' - the
shades are going \code{scale} of the way to white/black.
}
\details{
If \code{col} is provided and  \code{reps} is scalar it is recycled to the
length of \code{col}, otherwise it is filled with ones should
it be shorter than \code{col}.

The shades can be constructed either towards white, black, or, symmetrically.
This is determined by setting \code{direction} to 'bright', 'dark', or 'all',
respectively. The latter is the default.
When \code{gscale} is TRUE, all shade steps are the same, independent
of the individual repetitions, otherwise each shade range covers the range
given by \code{scale}.
}
\examples{
color.shades(3, c('red', 'blue'))
show.shades(color.shades(2:4, c('red', 'green', 'blue')))
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
