% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{ColorShadeRamp}
\alias{ColorShadeRamp}
\title{Create a color ramp function with shades of input}
\usage{
ColorShadeRamp(col, space = c("Lab", "rgb"), ...)
}
\arguments{
\item{col}{base color}

\item{space}{'Lab' or 'rgb', passed on to \code{\link[grDevices]{colorRamp}}}

\item{...}{also  passed on to \code{\link[grDevices]{colorRamp}}}
}
\value{
a function mapping [0,1] to (black ... \code{col} ... white)
}
\description{
The shades will range from black over \code{col} to white.
for symmetry the color should be a color with full saturation
}
\examples{
csr_red <- ColorShadeRamp('red')
show.colmat(t(csr_red(0:6/6)), width = 7)
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
