% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{concat.tbl.list}
\alias{concat.tbl.list}
\title{Concatenate a list of tables}
\usage{
concat.tbl.list(tbl.lst, idx = FALSE, ...)
}
\arguments{
\item{tbl.lst}{list of tables}

\item{idx}{whether or not to include "index" information}

\item{...}{ignored}
}
\value{
matrix corresponding to the \code{cbind}-concatenation of \code{tbl.lst}
}
\description{
Tables are
The individual tables---assumed to be output from \code{table}---in
\code{tbl.list} are split into their name and value part which are combined
to 2-row matrices via \code{rbind}. Those matrices are in turn concatenated
via \code{cbind}.
}
\details{
Parameter \code{idx} indicates whether or not to include "index" information,
i.e., indicate which table an entry came from (off by default)

Names of the sub-tables are assumed to be numeric as is the case for the table
lists
}
\examples{
table.list <- list(table(sample(1:3, 10, TRUE)), table(sample(7:10, 10, TRUE)))
concat.tbl.list(table.list)
concat.tbl.list(table.list, idx = TRUE)
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
