% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{cutshades}
\alias{cutshades}
\title{prepare shades and reorder to match cut}
\usage{
cutshades(cut, col.fun = default.rgb, ...)
}
\arguments{
\item{cut}{vector of cluster assignments}

\item{col.fun}{function mapping a number to that many colors from a palette}

\item{...}{passed to \code{\link{color.shades}}}
}
\value{
color shades reordered (see details)
}
\description{
For the given cut \code{cut} analyze the internal structure of each cluster
and use that as basis for a shading through \code{\link{color.shades}}. Then
reorder the shading to arrange according to cluster members.
}
\examples{
set.seed(42);
(cs <- cutshades(ct <-  sample(6, 15, TRUE)))
op <- par(mfrow=2:1)
show.colmat(cs, width=15)
show.shades(cs)
par(op)
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
