% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{default.cuts}
\alias{default.cuts}
\title{default cuts}
\usage{
default.cuts(n, ...)
}
\arguments{
\item{n}{number of leaves in dendrogram}

\item{...}{ignored}
}
\value{
nice cuts
}
\description{
find suitable cuts for \code{n} leaves
}
\details{
Internal, used with \code{\link{dummy.tree}}

For a square \code{n} the cuts are the multiples of \eqn{\sqrt{n}}{sqrt(n)}
up to \code{n}.
}
\examples{
default.cuts(9)
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
