% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{default.hcl}
\alias{default.hcl}
\title{HCL-based rainbow}
\usage{
default.hcl(n = 12, ...)
}
\arguments{
\item{n}{number of colors}

\item{...}{passed to rainbow_hcl}
}
\value{
rainbow colors
}
\description{
HCL-based rainbow
}
\examples{
default.hcl()
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
