% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{default.lab}
\alias{default.lab}
\title{Lab-based rainbow}
\usage{
default.lab(n = 12, ...)
}
\arguments{
\item{n}{number of colors}

\item{...}{passed to rainbow_lab}
}
\value{
rainbow colors
}
\description{
Lab-based rainbow
}
\examples{
default.lab()
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
