% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{default.rgb}
\alias{default.rgb}
\title{RGB-based rainbow}
\usage{
default.rgb(n = 12, ...)
}
\arguments{
\item{n}{number of colors}

\item{...}{passed to rainbow_rgb}
}
\value{
rainbow colors
}
\description{
RGB-based rainbow
}
\examples{
default.rgb()
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
