% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{dend.with.cuts}
\alias{dend.with.cuts}
\title{show dendrogram with cut levels overlaid}
\usage{
dend.with.cuts(hc.or.dt, cuts = NULL, cut.col = "#ffa050", ...)
}
\arguments{
\item{hc.or.dt}{output of \code{\link{dummy.tree}} or \code{link[stats]{hclust}}}

\item{cuts}{cut levels to show (cluster numbers) (optional, see description)}

\item{cut.col}{color to use for cut lines ("#ffa050" which is a light orange)}

\item{...}{passed to \code{\link{abline}} (\code{col}, \code{lty}, ...)}
}
\value{
invisible the cut levels corresponding to \code{cuts}
}
\description{
Mainly for debugging
}
\details{

}
\examples{
dt <- dummy.tree()
dend.with.cuts(dt)
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
