% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{dummy.tree}
\alias{dummy.tree}
\title{dummy tree data for examples}
\usage{
dummy.tree(n = 9, cuts = default.cuts(n), ..., seed = 1234)
}
\arguments{
\item{n}{number of classes [9]}

\item{cuts}{vector of cut numbers}

\item{...}{ignored}

\item{seed}{seed for randoom numbers to make code reproducible}
}
\value{
cutree result
}
\description{
The number shoud be reasonably small to keep the example results clear.
}
\details{
The \code{cut} vector specifies the number of branches in the dendrogram at
which the cuts should be made.
}
\examples{
dummy.tree()
dummy.tree(25)
dend.with.cuts(dummy.tree(25))
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
