% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{expand.colmat}
\alias{expand.colmat}
\title{expand color matrix}
\usage{
expand.colmat(cm, length = 256, fill.col = "#000000", ...)
}
\arguments{
\item{cm}{color matrix (or valid input to \code{\link[grDevices]{col2rgb}})}

\item{length}{desired number of color entries}

\item{fill.col}{color to fill matrix with ('black')}

\item{...}{ignored}
}
\value{
expanded color matrix
}
\description{
Add entries with \code{fill.col} to reach \code{length} colors in
color matrix.
}
\details{
Should \code{cm} have more entries than specified by \code{length},
the input martix is returned unchanged
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
