% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{\%&\%}
\alias{\%&\%}
\title{binary bitwise AND}
\usage{
a \%&\% b
}
\arguments{
\item{a}{boolean or numeric (integer)}

\item{b}{boolean or numeric (integer)}
}
\value{
bitwise AND of \code{a} and \code{b}
}
\description{
Wrapper for \code{\link[base]{bitwAnd}}, simply providing a binary interface.
}
\details{
While \code{a} or \code{b} can be any type of atomic variable, results may be
difficult to interpret for those other than logical or integer
}
\examples{
1 \%&\% 3
pi \%&\% 2
0.5 \%&\% 0.75
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
