% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{hri.plot}
\alias{hri.plot}
\title{hue range init plot}
\usage{
hri.plot(hues, y = 0.5, yoff = 0.1, add = FALSE, limits = NULL, ...)
}
\arguments{
\item{hues}{hue seeds, passed to \code{\link{hue.range.init}}}

\item{y}{y-level to show hue range line on (0..1)}

\item{yoff}{y-offset for start/end markers}

\item{add}{add to existing plot?}

\item{limits}{passed to  \code{\link{hue.range.init}}}

\item{...}{passed to \code{\link{hue.range.init}} (\code{limits}, \code{symmetric})}
}
\value{
new hue range (invisibly)
}
\description{
for internal testing
}
\details{
Show hue ranges created by \code{\link{hue.range.init}}.
}
\examples{
clusterLUTs:::hri.plot(1:5/6)
hues  <- c(.2, .4, .7, 5/6)
clusterLUTs:::hri.plot(hues,                                  yo = 0.05, y = 0.1)
clusterLUTs:::hri.plot(hues,                     symm = TRUE, yo = 0.05, y = 0.3, add = TRUE)
clusterLUTs:::hri.plot(hues, limits = c(.3, .8),              yo = 0.05, y = 0.5, add = TRUE)
clusterLUTs:::hri.plot(hues, limits = c(.3, .8), symm = TRUE, yo = 0.05, y = 0.7, add = TRUE)
}
