% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{hue.range.colors}
\alias{hue.range.colors}
\title{convert hue range matrix to corresponding colors}
\usage{
hue.range.colors(hr, extractfun = mean, only.hues = FALSE, ...)
}
\arguments{
\item{hr}{hue range matrix or a list thereof}

\item{extractfun}{function to calculate effective hue from hue range}

\item{only.hues}{when set, return (scalar) hue, otherwise convert to RGB
vector with \code{link[grDevices]{hsv}}}

\item{...}{passed to \code{\link[grDevices]{hsv}}}
}
\value{
color (vector) or hue(s) depending on \code{only.hues}
}
\description{
A hue range matrix is a \eqn{2\times N} matrix where each column defines a
range in hue space [0 .. 1] - if any hue value is found to be larger than one
the whole matrix is divided by 255.
}
\examples{
hr <- matrix(c(0.0, 0.6,
               0.2, 0.7), nrow = 2, byrow = TRUE)
hue.range.colors(hr)
hue.range.colors(hr, only.hues = TRUE)
show.colmat(col2rgb(hue.range.colors(hr)))
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
