% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{hue.range.init}
\alias{hue.range.init}
\title{hue range initialization}
\usage{
hue.range.init(hues = 5/12, blank = 0.1, limits = NULL, symmetric = FALSE, ...)
}
\arguments{
\item{hues}{initial "centers"}

\item{blank}{fraction of full range (given by \code{limits}) not to assign}

\item{limits}{define full range of available hues}

\item{symmetric}{range symmetric around \code{hues} (where possible)}

\item{...}{ignored}
}
\value{
new hue range
}
\description{
Initialize a hue range "seeded" by \code{hues}. It is attempted to
allocate the range given by \code{limits} sparing a fraction of
\code{blank} for the breaks. If \code{symmetric} is set, the hue
ranges are centered around the seed values (except when at the
limits) leaving more of the hue space unallocated.
}
\details{
The values in \code{hues} are trimmed to [0..1] and duplicates are removed.
}
\author{
Puetz
}
