% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{hue.range.lines}
\alias{hue.range.lines}
\title{plot hue range}
\usage{
hue.range.lines(hue.range, y = 1, add = FALSE, ..., col = "red")
}
\arguments{
\item{hue.range}{hue range or list thereof}

\item{y}{y-level at which to plot (scalar or vector of length of \code{hue.range})}

\item{add}{when set, add to existing plot, otherwise start a new one}

\item{...}{passed on to \code{\link{segments}}}

\item{col}{consume any color that may be provided}
}
\value{
none
}
\description{
for now only horizontal
}
\examples{
hue.range.lines(matrix(0:5/6, nr = 2))
}
\seealso{
\code{\link{hue.range.colors}} for an unordered display.
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
