% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{init.blank.plot}
\alias{init.blank.plot}
\title{blank plot}
\usage{
init.blank.plot(...)
}
\arguments{
\item{...}{goes to plot}
}
\value{
none
}
\description{
just an empty plot [0,1]x[0,1] with nothing on it (no axes, no labels)
}
\examples{
init.blank.plot()
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
