% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{init.huerange.plot}
\alias{init.huerange.plot}
\title{initialize hue range plot}
\usage{
init.huerange.plot(...)
}
\arguments{
\item{...}{ignored}
}
\value{
none
}
\description{
empty "frame" (just x-axis) with x- and y-range 0:1
}
\examples{
init.huerange.plot()
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
