% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{mat2list}
\alias{mat2list}
\title{matrix to list}
\usage{
mat2list(mat, which = 2, ...)
}
\arguments{
\item{mat}{input matrix}

\item{which}{1 for row, 2 for column vectors}

\item{...}{ignored}
}
\value{
list with matrix split in vectors
}
\description{
Convert matrix \code{mat} to a list of either column vectors (for \code{which == 2})
or row vectors  (\code{which == 1}).
}
\examples{
m <- matrix(1:6, ncol = 2)
mat2list(m)
mat2list(m, 1)
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
