% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{rainbow_lab}
\alias{rainbow_lab}
\title{rainbow via Lab-based colorRamp}
\usage{
rainbow_lab(n, rlf = rainbow_lab_ramp(), ...)
}
\arguments{
\item{n}{number of colors to return}

\item{rlf}{rainbow Lab function (\code{\link{rainbow_lab_ramp}})}

\item{...}{passed to \code{rlf}}
}
\value{
color matrix
}
\description{
Create a Lab-based rainbow color matrix.
}
\examples{
rainbow_lab(12)
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
