% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{rainbow_lab_ramp}
\alias{rainbow_lab_ramp}
\title{LAB-based Rainbow colorramp function}
\usage{
rainbow_lab_ramp(full = FALSE)
}
\arguments{
\item{full}{whether to go all around to red or only to purple}
}
\value{
color ramp function
}
\description{
Calculating the rainbow colors in LAB color space gives slightly better distinction
between neighboring colors.
}
\details{
For normal usage \code{full=FALSE} should yield good results (hues from red
to purple). If, however, a correspondence to hue is desired,
 \code{full=TRUE} should be used to cover
the full hue range from [0..1].
}
\examples{
rainbow_lab_ramp()(0)       # red
rainbow_lab_ramp()(1)       # purple
rainbow_lab_ramp(TRUE)(1)   # red
show.lut(t(rainbow_lab_ramp()(0:11/11)))
show.lut(t(rainbow_lab_ramp(TRUE)(0:11/11)))
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
