% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{randomize.cutree}
\alias{randomize.cutree}
\title{randomize cluster assignments}
\usage{
randomize.cutree(cutree, ...)
}
\arguments{
\item{cutree}{result of  \code{\link[stats]{cutree}}}

\item{...}{ignored}
}
\value{
reordered matrix
}
\description{
In R's implementation of \code{\link[stats]{cutree}} the first element is
always assigned to cluster 1, unlike what Matlab produces. This function
attemps to make a \code{\link[stats]{cutree}} result more similar to it's
Matlab counterpart (or what \code{\link{read.tree}} produces of it).
}
\details{
All the columns where the number of levels does not equal the number of rows
(elements) is reordered (and it is considered a problem if no full column is
found).
}
\examples{
ct <- dummy.tree()
cbind(ct,randomize.cutree(ct))
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
