% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{read.tree}
\alias{read.tree}
\title{read cluster tree information as obtained through Matlab}
\usage{
read.tree(
  file = system.file("extdata/sbm_1_145_0_atlas.mat", package = "clusterLUTs"),
  path = ".",
  add.full = TRUE,
  ...
)
}
\arguments{
\item{file}{Matlab hclust result file (Matlab *.mat-file)}

\item{path}{directory where to find file}

\item{add.full}{add a fully specified column if there is none?}

\item{...}{ignored}
}
\value{
a matrix with just the cluster assignments
}
\description{
The mat-file is expected to have one component "cluster.info" which, in turn
contains a list of lists. Each of the latter is a \eqn{n\times 1}{nx1} matrix with cluster
assignments (the number of clusters at each cut is given by the number of unique
entries)
}
\details{
Each of these matrices (vectors) is forms one column in the output matrix as
would be obtained by \code{\link[stats]{cutree}}. The matrix is given column
names indicating the number of clusters for that cut.
}
\examples{
   tree <- read.tree()
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
