% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{readlut}
\alias{readlut}
\title{read lookup table}
\usage{
readlut(file, length = file.size(file)\%/\%3, ...)
}
\arguments{
\item{file}{lookup table file}

\item{length}{number of color entries to read}

\item{...}{ignored}
}
\value{
color matrix (see description)
}
\description{
The file will be read as binary with three (unsigned) bytes per color and
return a color matrix with columns corresponding to the color entries (RGB).
}
\details{
If \code{length} is specified, read only that many entries, otherwise all.
Some consistency checks on the file size are performed.
}
\examples{
\dontrun{
   lut <-  readlut('test.lut')
}
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
