% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{reidx.cut}
\alias{reidx.cut}
\title{re-index cut(ree) vector}
\usage{
reidx.cut(cut)
}
\arguments{
\item{cut}{cut from hierarchical clustering (through \code{link[stats]{cutree}})}
}
\value{
rank vector with names to indicate level and order within ties
}
\description{
Internal rank increasing within ties so that after reordering the order of ties
can be preserved
}
\details{
the output vector is named with (original level + position as decimal fraction) with ties for
that level, see example. this can handles multiplicities of up to 999 which
should suffice for most practical cases.

This is useful to reorder, e.g., a color LUT from \code{\link{color.shades}}
according to levels.
}
\examples{
ct <- c(4, 1, 2, 3, 1, 1, 1, 1, 4, 1, 1, 3, 1, 3)
reidx.cut(ct)
## 4.1 1.1 2.1 3.1 1.2 1.3 1.4 1.5 4.2 1.6 1.7 3.2 1.8 3.3
##  13   1   9  10   2   3   4   5  14   6   7  11   8  12
## the first occurence of level 1 is named 1001 (see details)
cs <- color.shades(table(ct))
rocs <- cs[, reidx.cut(ct)]
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
