% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{show.brain.lut}
\alias{show.brain.lut}
\title{show brain lookup table}
\usage{
show.brain.lut(
  n,
  LUTdir,
  lut = NULL,
  clusters = read.tree(),
  verbose = getOption("verbose"),
  test = FALSE,
  ...
)
}
\arguments{
\item{n}{number of cuts the table should have}

\item{LUTdir}{directory where to look for LUT files}

\item{lut}{alternative LUT, see description}

\item{clusters}{matrix with cluster assignments (1 column per cut)}

\item{verbose}{provide feedback?}

\item{test}{set to give specific return strings for tests}

\item{...}{ignored}
}
\value{
LUT used (invisibly)
}
\description{
show one of the lookup tables created through \code{\link{treeluts}}
}
\details{
The LUT can, alternatively, be given through \code{lut}: either a compatible
matrix (3 rows or columns) or a complete file path
}
\section{Testing}{

When \code{test} is set, specific strings are returned for use with
\code{\link[testthat]{expect_equal}}.
}

\examples{
\dontrun{
  show.brain.lut(10, 'luts')
}
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
