% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{show.colmat}
\alias{show.colmat}
\title{show color matrix}
\usage{
show.colmat(cv, width = NULL, show.idx = FALSE, ...)
}
\arguments{
\item{cv}{color vector (actually \eqn{3\times n}{3xn} matrix with one column per color (RGB))}

\item{width}{how many color tiles across (~\eqn{\sqrt{n}}{sqrt(n)} where \eqn{n}
is the length of the color vector)}

\item{show.idx}{add index to display}

\item{...}{passed to \code{\link[graphics]{rect}}}
}
\value{
none
}
\description{
Show colors represented in color vector \code{cv} as matrix of colored boxes.
}
\details{
By setting \code{show.idx} to TRUE, the index in the color vector is shown in
the center of the color tiles. Assigning a vector the same length as \code{cv},
the elements of that vector will be shown rather than the index.
}
\examples{
op <- par(mfrow=c(2,1))
show.colmat(rainbow(12, end = 0.8))
show.colmat(rainbow(12, end = 0.8), width = 12)
par(op)
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
