% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{show.cut}
\alias{show.cut}
\title{show cuts}
\usage{
show.cut(cut)
}
\arguments{
\item{cut}{vector with cluster assignments}
}
\value{
none
}
\description{
show how levels of cut relate to individual features
}
\examples{
show.cut(dummy.tree()[[1]])
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
