% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{show.hue.range}
\alias{show.hue.range}
\title{Show hue range colors}
\usage{
show.hue.range(
  hr,
  extractfun = mean,
  spacing = 0.1,
  y = NULL,
  width = NULL,
  v.scale = 1,
  show.idx = FALSE,
  ...,
  wait = FALSE,
  width.x = width,
  width.y = width
)
}
\arguments{
\item{hr}{hue range or list thereof}

\item{extractfun}{function to assign hue to range (\code{\link[base]{mean}})}

\item{spacing}{relative spacing between hue lines}

\item{y}{position along y-axis (mainly for single hue range)}

\item{width}{effective width of rectangles along axes}

\item{v.scale}{relative vertical scaling to adjust rectangles' height}

\item{show.idx}{add index (if boolean) or specified values (if char, numeric)}

\item{...}{ignored}

\item{wait}{for internal use, should not be set by caller}

\item{width.x}{width along x-axis}

\item{width.y}{width along y-axis}
}
\value{
none
}
\description{
Somehere between \code{\link{show.colmat}} and \code{\link{hue.range.lines}},
shows the unordered colors of one in a single row of rectangles.
}
\details{
When given a list of hue ranges, those lines are stacked on top of one
another.
}
\examples{
show.hue.range(matrix(0:5/6, nr = 2))
}
\seealso{
\code{\link{hue.range.lines}} or \code{\link{tree.ranges.plot}} for
ordered display.
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
