% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{show.lut}
\alias{show.lut}
\title{show LUT ([color] lookup table)}
\usage{
show.lut(lut, colorspace = c("RGB", "HSV"), ...)
}
\arguments{
\item{lut}{lookup table}

\item{colorspace}{should be either 'RGB' or 'HSV'}

\item{...}{ignored}
}
\value{
none
}
\description{
Give a visual representation of the colors represented in a LUT.
}
\details{
First a color band showing the colors is presented, below the RGB components are plotted
}
\examples{
show.lut(rainbow(12, end = 11/12, v  = 0.9))
}
