% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{show.shades}
\alias{show.shades}
\title{Show shading in histogram-style plot}
\usage{
show.shades(shades, orient = c("horizontal", "vertical"), r = 0.45, ...)
}
\arguments{
\item{shades}{shades as produced by \code{\link{color.shades}}}

\item{orient}{orientation for main colors}

\item{r}{"radius" for color boxes, should be less than 0.5}

\item{...}{ignored}
}
\value{
none
}
\description{
The color matrix \code{shades} is expected to have an attribute 'reps' indicating
the number of shades for each main color (and, thus, the number of main colors.)
}
\details{
The "base" colors are shown along the axis indicated by \code{orient}, the shades
are stacked perpendicular to that. If the shades were created with \code{dir = 'all'},
the shade stacks are aligned on the base colors
}
\examples{
rainbow.shades <- color.shades(sample(1:10, 20, TRUE))
show.shades(rainbow.shades)
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
