% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{show.sub.tables}
\alias{show.sub.tables}
\title{show sub tables}
\usage{
show.sub.tables(st, which = NULL, ..., wait = FALSE)
}
\arguments{
\item{st}{subtables or list thereof (or result of  \code{\link{tree.ranges}})}

\item{which}{optional subset of the subtables}

\item{...}{ignored}

\item{wait}{for internal use, should not be set by caller}
}
\value{
string representation of sub tables
}
\description{
In the context of this package, subtable refers to either a single table or a
list of tables which represents chunks of a higher level table (thus subtable).
They are used to trace clustering trees.
}
\details{
If \code{st} is a list of lists, \code{\link{show.sub.tables}} will be called
recursively. For convenience, it is also possible to directly provide the output
of \code{\link{tree.ranges}}, where the attribute \code{sub.tables} is extracted
and used.

At this point the table relation across levels is not directly shown, it has
to be infered from the labels
}
\examples{
show.sub.tables(attr(tree.ranges(dummy.tree()), "sub.tables"))
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
