% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{subtables}
\alias{subtables}
\title{subtables}
\usage{
subtables(tr, ...)
}
\arguments{
\item{tr}{arbitrary object}

\item{...}{ignored}
}
\value{
'sub.tables' attribute of \code{tr}, if present, failure otherwise
}
\description{
accessor function to 'sub.tables' attribute of \code{tree.ranges}.
}
\details{
Return 'sub.tables' attribute of \code{tr}, if present, failure otherwise.
}
\examples{
subtables(tree.ranges(dummy.tree()))
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
