% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{tbl2mat}
\alias{tbl2mat}
\title{table to matrix}
\usage{
tbl2mat(tbl, uplink = NULL, ...)
}
\arguments{
\item{tbl}{output of \code{\link[base]{table}}}

\item{uplink}{extra row(s) to be included (see description)}

\item{...}{ignored}
}
\value{
a 2-row matrix representing the table.
}
\description{
convert table to matrix
}
\details{
Create a matrix with names and values of the input table forming a row each.

It is assumed that \code{tbl}'s names are all numeric

Parameter \code{uplink} allows inclusion of additional rows below the
defaults rows 'idx' and 'n'. If provided as matrix, \code{uplink}
should have rownames.
}
\examples{
tbl2mat(table(1:3))
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
