% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{tree.ranges.plot}
\alias{tree.ranges.plot}
\title{plot hues for a \code{\link{tree.ranges}} result}
\usage{
tree.ranges.plot(
  tr,
  add = FALSE,
  show.tree = FALSE,
  ts.col = "lightgrey",
  ts.lty = 1,
  ts.lwd = 1,
  ...
)
}
\arguments{
\item{tr}{output of \code{\link{tree.ranges}}}

\item{add}{add to existing plot or start new one}

\item{show.tree}{show lines indicating the tree structure}

\item{ts.col}{tree segment color}

\item{ts.lty}{tree segment line type}

\item{ts.lwd}{tree segment line width}

\item{...}{passed to \code{\link{hue.range.lines}}}
}
\value{
none
}
\description{
plot hues for a \code{\link{tree.ranges}} result
}
\examples{
tr <- tree.ranges(dummy.tree())
tree.ranges.plot(tr, show.tree = TRUE)
}
\seealso{
\code{\link{hue.range.colors}} for an unordered display
}
