% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{tree.ranges}
\alias{tree.ranges}
\title{full hue range set for cutree}
\usage{
tree.ranges(cuts, blank = 0.1, init = NULL, ...)
}
\arguments{
\item{cuts}{cluster assignment result of \code{\link[stats]{cutree}} (or equivalent)}

\item{blank}{"spacing" between hues ranges when splitting (see description)}

\item{init}{provide initial hues - not yet implemented}

\item{...}{passed to \code{\link{hue.range.split}}}
}
\value{
list of related hue ranges corresponding to the cut levels
}
\description{
Create a set of hue ranges related through the levels of \code{cuts}.
}
\details{
When \code{blank} is non-zero, that fraction of a hue range is left unassigned
when splitting for the next level. This should help for better distingushable
colors in the resulting hue ranges.
}
\examples{
tree.ranges(dummy.tree())
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
