% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{treeluts}
\alias{treeluts}
\title{LUT generation for hierarchical clustering tree}
\usage{
treeluts(
  clusters = read.tree(),
  outdir = "luts",
  basename = "lut",
  lut.length = 256,
  verbose = getOption("verbose"),
  ...
)
}
\arguments{
\item{clusters}{matrix with cluster assignments, one column per cut}

\item{outdir}{where to write the LUTs to}

\item{basename}{prefix for LUT names ['lut']}

\item{lut.length}{length of the LUT (number of entries [256])}

\item{verbose}{set to TRUE to see the LUT filenames}

\item{...}{ignored}
}
\value{
a vector with the number of color entries per LUT generated (invisibly)
}
\description{
{
Main function of this package: \code{clusters} defines the cluster assignments
per cut. The number of clusters
is taken from the unique values per column (need not be regular).
}
}
\details{

}
\examples{
set.seed(42)
treeluts(cbind(sample(1:5,20,TRUE),1:20), tempdir(), "exmpl", verbose=TRUE)
\dontrun{
 treeluts(outdir = path.expand("~/Work/4philipp/BrainLUTs/luts"),  verbose=TRUE)
}
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
