% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{vec2hsv}
\alias{vec2hsv}
\title{Vector to HSV conversion}
\usage{
vec2hsv(v, ...)
}
\arguments{
\item{v}{a 3-element vector with the values for \code{h}, \code{s}, and \code{v}}

\item{...}{ignored}
}
\value{
an RGB vector (1-column color matrix) representing the color
}
\description{
Wrapper for \code{\link[grDevices]{hsv}} that allows using it with a 3-element vector
\code{[h, s, v]}.
}
\details{
Rather than returning a hex representation as done in \code{\link[grDevices]{hsv}}, a
conversion to RGB is performed before returning.
}
\examples{
# convert RGB colors to HSV and back
show.lut(apply(apply(col2rgb(rainbow(6, e=5/6)), 2, rgb2hsv), 2, vec2hsv))
}
