% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{vec2rgb}
\alias{vec2rgb}
\title{convert 3-element RGB-vector to color}
\usage{
vec2rgb(v, maxColorValue = ifelse(any(v > 1), 255, 1), ...)
}
\arguments{
\item{v}{3- or 4-element RGB-vector}

\item{maxColorValue}{passed on to \code{\link{rgb}}}

\item{...}{ignored}
}
\value{
corresponding color
}
\description{
Wrapper for \code{\link{rgb}} that allows using it with a 3-element vector
\code{[r, g, b]}.
}
\details{
If \code{maxColorValue} is not provided the appropriate value is guessed from
the values of \code{v}. If any is larger than 1, 255 is assumed. In most cases
that should be fine.
}
\examples{
vec2rgb(1:3*80)
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
