% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterluts.R
\name{vlevels}
\alias{vlevels}
\title{vector levels}
\usage{
vlevels(v)
}
\arguments{
\item{v}{vector}
}
\value{
number of unique elements in \code{v}
}
\description{
Count the number of unique entries in vector (usually an index vector with
integer elements)
}
\examples{
v <- c(1:4, 1:3, 1:2, 1)
vlevels(v)
table(v)
}
\author{
Benno Pütz \email{puetz@psych.mpg.de}
}
