% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{writelut}
\alias{writelut}
\title{write LUT}
\usage{
writelut(x, filename)
}
\arguments{
\item{x}{vector of 768 bytes (treated as unsigned integer)}

\item{filename}{name for output file}
}
\value{
0 on success
}
\description{
Write RGB color lookup table (LUT) of 3\eqn{\times}{x}256 bytes. 
Implemented with Rcpp to allow writing bytes.
}
